everything should function with data? hopefully? maybe? if you just believe

+++ OrientationApi/OrientationApi/Models/Product.cs
@@ -9,6 +9,6 @@ namespace OrientationApi.Models
     {
         public int ProductId { get; set; }
         public string ProductName { get; set; }
-        public double ProductPrice { get; set; }
+        public int ProductPrice { get; set; }
     }
 }
